[Server] Multi-Threaded Processing of Messages
Changed the layouts of Query Handler and Registration Module to go for
threaded operation. The idea is that each incoming message should be
processed in its own thread and all it's work should be finished in that
thread; so that we minimize the possibility of blocking the main server
thread.