Fixed Models. Finished Registration and de-Reg
->Fixed Models a bit.
->Changed the sqlite memory connection to connect to a file. It creates
the file on its own if it does not exist. If testing, please make sure
to manually call the "connect" method of the Models file once, so that
the required tables are created. The file is created in the "src" folder
if project is run from Eclipse.
-> Finished and tested Registration/De Registration at local machine.
Yet to Push on the server.